# Supabase-Gateway/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SahatGhar Telemedicine Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List all appointments
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: doctor_id
          in: query
          schema:
            type: string
        - name: patient_id
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: date_from
          in: query
          schema:
            type: string
        - name: date_to
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentListResponse"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Appointment detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"

    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
...
  /admin-users:
    get:
      operationId: listAdminUsers
      tags: [admin-users]
      summary: List admin accounts
      responses:
        "200":
          description: Admin users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"

    post:
      operationId: createAdminUser
      tags: [admin-users]
      summary: Create an admin account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin-users/{id}/status:
    patch:
      operationId: updateAdminUserStatus
      tags: [admin-users]
      summary: Suspend or activate an admin user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated admin user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
components:
  schemas:
...
    Appointment:
      type: object
      required: [id, patient_name, doctor_name, date_time, type, city, status, amount]
      properties:
        id:
          type: string
        patient_name:
          type: string
        patient_age:
          type: ["integer", "null"]
        patient_gender:
          type: ["string", "null"]
        doctor_name:
          type: string
        doctor_specialty:
          type: ["string", "null"]
        date_time:
          type: string
        type:
          type: string
        city:
          type: string
        status:
          type: string
        amount:
          type: number
        payment_method:
          type: ["string", "null"]
        payment_status:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    AppointmentListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AppointmentUpdate:
      type: object
      properties:
        status:
          type: string
        notes:
          type: string
...
    AdminUser:
      type: object
      required: [id, name, email, role, status, last_active]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        status:
          type: string
        last_active:
          type: string
        avatar_url:
          type: ["string", "null"]
        failed_attempts:
          type: ["integer", "null"]
    AdminUserInput:
      type: object
      required: [name, email, role]
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
    StatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
        reason:
          type: ["string", "null"]